Sketch workflow engine abstraction
