+++ crates/rt/src/utils.rs
@@ -1,10 +1,8 @@
 //! Misc utils.
 
-use anyhow::Context as _;
 use futures::{future::Aborted, stream::BoxStream, Stream, StreamExt};
 use ouroboros::self_referencing;
 use serde::{Deserialize, Serialize};
-use wasmtime::{AsContext, AsContextMut, Memory, StoreContextMut};
 
 use std::{
     fmt,
@@ -13,19 +11,7 @@ use std::{
     task::{Context, Poll},
 };
 
-use crate::data::WorkflowData;
-use tardigrade::{abi::AllocateBytes, task::JoinError};
-
-pub(crate) fn debug_result<T, E>(result: &Result<T, E>)
-where
-    T: fmt::Debug,
-    E: fmt::Debug + fmt::Display,
-{
-    match result {
-        Ok(_) => tracing::debug!(result = ?result),
-        Err(err) => tracing::warn!(result.err = %err, "erroneous result"),
-    }
-}
+use tardigrade::task::JoinError;
 
 /// Thin wrapper around `Vec<u8>`.
 #[derive(Clone, PartialEq, Serialize, Deserialize)]
@@ -57,64 +43,6 @@ impl From<Message> for Vec<u8> {
     fn from(message: Message) -> Self {
         message.0
     }
-}
-
-pub(crate) struct WasmAllocator<'ctx, 'a>(StoreContextMut<'ctx, WorkflowData<'a>>);
-
-impl fmt::Debug for WasmAllocator<'_, '_> {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter.debug_tuple("WasmAllocator").field(&"_").finish()
-    }
-}
-
-impl<'ctx, 'a> WasmAllocator<'ctx, 'a> {
-    pub fn new(ctx: StoreContextMut<'ctx, WorkflowData<'a>>) -> Self {
-        Self(ctx)
-    }
-}
-
-impl AllocateBytes for WasmAllocator<'_, '_> {
-    type Error = anyhow::Error;
-
-    #[tracing::instrument(level = "trace", skip_all, ret, err, fields(bytes.len = bytes.len()))]
-    fn copy_to_wasm(&mut self, bytes: &[u8]) -> anyhow::Result<(u32, u32)> {
-        let bytes_len =
-            u32::try_from(bytes.len()).context("integer overflow for message length")?;
-        let exports = self.0.data().exports();
-        let ptr = exports.alloc_bytes(self.0.as_context_mut(), bytes_len)?;
-
-        let host_ptr = usize::try_from(ptr).unwrap();
-        let memory = self.0.data_mut().exports().memory;
-        memory
-            .write(&mut self.0, host_ptr, bytes)
-            .context("cannot write to WASM memory")?;
-        Ok((ptr, bytes_len))
-    }
-}
-
-pub(crate) fn copy_bytes_from_wasm(
-    ctx: impl AsContext,
-    memory: &Memory,
-    ptr: u32,
-    len: u32,
-) -> anyhow::Result<Vec<u8>> {
-    let ptr = usize::try_from(ptr).unwrap();
-    let len = usize::try_from(len).unwrap();
-    let mut buffer = vec![0_u8; len];
-    memory
-        .read(ctx, ptr, &mut buffer)
-        .context("error copying memory from WASM")?;
-    Ok(buffer)
-}
-
-pub(crate) fn copy_string_from_wasm(
-    ctx: impl AsContext,
-    memory: &Memory,
-    ptr: u32,
-    len: u32,
-) -> anyhow::Result<String> {
-    let buffer = copy_bytes_from_wasm(ctx, memory, ptr, len)?;
-    String::from_utf8(buffer).map_err(From::from)
 }
 
 #[cfg(test)]
