+++ crates/rt/src/workflow/persistence.rs
@@ -1,18 +1,18 @@
 //! Workflow persistence.
 
-use anyhow::{ensure, Context};
+use anyhow::Context;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
-use wasmtime::Store;
 
-use std::{fmt, task::Poll};
+use std::task::Poll;
 
 use crate::{
     data::{
         Channels, ChildWorkflow, ConsumeError, PersistError, PersistedWorkflowData, ReceiverState,
-        Refs, SenderState, TaskState, TimerState, Wakers, WorkflowData,
+        SenderState, TaskState, TimerState, Wakers,
     },
-    module::{DataSection, Services, WorkflowSpawner},
+    engine::{CreateWorkflow, PersistWorkflow, RunWorkflow},
+    module::{Services, WorkflowSpawner},
     receipt::WakeUpCause,
     utils::Message,
     workflow::{ChannelIds, Workflow},
@@ -23,169 +23,46 @@ use tardigrade::{
     ChannelId, TaskId, TimerId, WorkflowId,
 };
 
-#[derive(Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum Memory {
-    Unstructured(#[serde(with = "serde_compress")] Vec<u8>),
-    Structured {
-        data_base: u32,
-        #[serde(with = "serde_compress")]
-        data_diff: Vec<u8>,
-        #[serde(with = "serde_compress")]
-        heap: Vec<u8>,
-    },
-}
-
-impl fmt::Debug for Memory {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Unstructured(bytes) => formatter
-                .debug_struct("Unstructured")
-                .field("len", &bytes.len())
-                .finish(),
-
-            Self::Structured {
-                data_base,
-                data_diff,
-                heap,
-            } => formatter
-                .debug_struct("Structured")
-                .field("data_base", data_base)
-                .field("data_diff_len", &data_diff.len())
-                .field("heap_len", &heap.len())
-                .finish(),
-        }
-    }
-}
-
-mod serde_compress {
-    use flate2::{
-        bufread::{DeflateDecoder, DeflateEncoder},
-        Compression,
-    };
-    use serde::{de::Error as _, ser::Error as _, Deserializer, Serializer};
-
-    use std::io::{BufReader, Read};
-
-    use crate::utils::serde_b64;
-
-    pub fn serialize<S: Serializer>(bytes: &[u8], serializer: S) -> Result<S::Ok, S::Error> {
-        let reader = BufReader::new(bytes);
-        let mut compressor = DeflateEncoder::new(reader, Compression::best());
-        let mut output = vec![];
-        compressor
-            .read_to_end(&mut output)
-            .map_err(|err| S::Error::custom(format!("cannot compress memory snapshot: {}", err)))?;
-
-        serde_b64::serialize(&output, serializer)
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let compressed: Vec<u8> = serde_b64::deserialize(deserializer)?;
-        let reader = BufReader::new(compressed.as_slice());
-        let mut output = vec![];
-        DeflateDecoder::new(reader)
-            .read_to_end(&mut output)
-            .map_err(|err| {
-                D::Error::custom(format!("cannot decompress memory snapshot: {}", err))
-            })?;
-        Ok(output)
-    }
-}
-
-impl Memory {
-    fn new(store: &Store<WorkflowData>, data_section: Option<&DataSection>) -> Self {
-        let exports = store.data().exports();
-        if let Some(section) = data_section {
-            let memory = exports.memory.data(store);
-            Self::Structured {
-                data_base: section.start(),
-                data_diff: section.create_diff(memory),
-                heap: memory[section.end()..].to_vec(),
-            }
-        } else {
-            Self::Unstructured(exports.memory.data(store).to_vec())
-        }
-    }
-
-    fn restore(self, workflow: &mut Workflow) -> anyhow::Result<()> {
-        match self {
-            Self::Unstructured(bytes) => {
-                workflow
-                    .copy_memory(0, &bytes)
-                    .context("failed restoring workflow memory")?;
-            }
-            Self::Structured {
-                data_base,
-                mut data_diff,
-                heap,
-            } => {
-                let section = workflow
-                    .data_section
-                    .as_deref()
-                    .ok_or_else(|| anyhow::anyhow!("expected data section"))?;
-                ensure!(
-                    section.start() == data_base && section.len() == data_diff.len(),
-                    "data section shape differs in the persisted workflow and workflow module"
-                );
-                let heap_start = section.end();
-
-                section.restore_from_diff(&mut data_diff);
-                workflow
-                    .copy_memory(data_base as usize, &data_diff)
-                    .context("failed restoring workflow data section")?;
-                workflow
-                    .copy_memory(heap_start, &heap)
-                    .context("failed restoring workflow heap")?;
-            }
-        }
-        Ok(())
-    }
-}
-
 /// Persisted version of a workflow containing the state of its external dependencies
 /// (channels and timers), and its linear WASM memory.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct PersistedWorkflow {
-    state: PersistedWorkflowData,
-    refs: Refs,
-    memory: Memory,
+    data: PersistedWorkflowData,
+    engine_data: serde_json::Value,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     args: Option<Message>,
 }
 
 impl PersistedWorkflow {
-    pub(super) fn new(mut workflow: Workflow<'_>) -> Result<Self, PersistError> {
-        workflow.store.data().check_persistence()?;
-        let refs = Refs::new(&mut workflow.store);
-        let memory = Memory::new(&workflow.store, workflow.data_section.as_deref());
-        let data = workflow.store.into_data();
+    pub(super) fn new<T>(workflow: &mut Workflow<T>) -> Result<Self, PersistError>
+    where
+        T: RunWorkflow + PersistWorkflow,
+    {
+        workflow.inner.data().check_persistence()?;
+        let engine_data = workflow.inner.persist();
+        let engine_data = serde_json::to_value(&engine_data).expect("cannot serialize engine data");
         Ok(Self {
-            state: data.persist(),
-            refs,
-            memory,
-            args: workflow.raw_args.clone(),
+            data: workflow.inner.data().persist(),
+            engine_data,
+            args: workflow.args.clone(),
         })
     }
 
     pub(crate) fn receivers(&self) -> impl Iterator<Item = (ChannelId, &ReceiverState)> + '_ {
-        self.state.receivers()
+        self.data.receivers()
     }
 
     pub(crate) fn receiver(&self, channel_id: ChannelId) -> Option<&ReceiverState> {
-        self.state.receiver(channel_id)
+        self.data.receiver(channel_id)
     }
 
     pub(crate) fn senders(&self) -> impl Iterator<Item = (ChannelId, &SenderState)> + '_ {
-        self.state.senders()
+        self.data.senders()
     }
 
     /// Returns information about channels defined in this workflow interface.
     pub fn channels(&self) -> Channels<'_> {
-        self.state.channels()
+        self.data.channels()
     }
 
     #[tracing::instrument(level = "debug", skip(self, message), err, fields(message.len = message.len()))]
@@ -194,49 +71,49 @@ impl PersistedWorkflow {
         channel_id: ChannelId,
         message: Vec<u8>,
     ) -> Result<(), ConsumeError> {
-        self.state.push_message_for_receiver(channel_id, message)
+        self.data.push_message_for_receiver(channel_id, message)
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
     pub(crate) fn drop_message_for_receiver(&mut self, channel_id: ChannelId) {
-        self.state.drop_message_for_receiver(channel_id);
+        self.data.drop_message_for_receiver(channel_id);
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
     pub(crate) fn close_receiver(&mut self, channel_id: ChannelId) {
-        self.state.close_receiver(channel_id);
+        self.data.close_receiver(channel_id);
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
     pub(crate) fn close_sender(&mut self, channel_id: ChannelId) {
-        self.state.close_sender(channel_id);
+        self.data.close_sender(channel_id);
     }
 
     /// Returns the current state of a task with the specified ID.
     pub fn task(&self, task_id: TaskId) -> Option<&TaskState> {
-        self.state.task(task_id)
+        self.data.task(task_id)
     }
 
     /// Returns the current state of the main task (the task that initializes the workflow),
     /// or `None` if the workflow is not initialized.
     pub fn main_task(&self) -> Option<&TaskState> {
-        self.state.main_task()
+        self.data.main_task()
     }
 
     /// Lists all tasks in this workflow.
     pub fn tasks(&self) -> impl Iterator<Item = (TaskId, &TaskState)> + '_ {
-        self.state.tasks()
+        self.data.tasks()
     }
 
     /// Enumerates child workflows.
     pub fn child_workflows(&self) -> impl Iterator<Item = (WorkflowId, ChildWorkflow<'_>)> + '_ {
-        self.state.child_workflows()
+        self.data.child_workflows()
     }
 
     /// Returns the local state of the child workflow with the specified ID, or `None`
     /// if a workflow with such ID was not spawned by this workflow.
     pub fn child_workflow(&self, id: WorkflowId) -> Option<ChildWorkflow<'_>> {
-        self.state.child_workflow(id)
+        self.data.child_workflow(id)
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
@@ -245,7 +122,7 @@ impl PersistedWorkflow {
         id: WorkflowId,
         result: Result<(), JoinError>,
     ) {
-        self.state.notify_on_child_completion(id, result);
+        self.data.notify_on_child_completion(id, result);
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
@@ -256,13 +133,13 @@ impl PersistedWorkflow {
         channels: &ChannelsConfig<ChannelId>,
         channel_ids: ChannelIds,
     ) {
-        self.state
+        self.data
             .notify_on_child_init(stub_id, id, channels, channel_ids);
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
     pub(crate) fn notify_on_child_spawn_error(&mut self, stub_id: WorkflowId, err: HostError) {
-        self.state.notify_on_child_spawn_error(stub_id, err);
+        self.data.notify_on_child_spawn_error(stub_id, err);
     }
 
     /// Checks whether the workflow is initialized.
@@ -272,37 +149,37 @@ impl PersistedWorkflow {
 
     /// Returns the result of executing this workflow, which is the output of its main task.
     pub fn result(&self) -> Poll<Result<(), &JoinError>> {
-        self.state.result()
+        self.data.result()
     }
 
     /// Aborts the workflow by changing the result of its main task.
     pub(crate) fn abort(&mut self) {
-        self.state.abort();
+        self.data.abort();
     }
 
     /// Returns the current time for the workflow.
     pub fn current_time(&self) -> DateTime<Utc> {
-        self.state.timers.last_known_time()
+        self.data.timers.last_known_time()
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
     pub(crate) fn set_current_time(&mut self, time: DateTime<Utc>) {
-        self.state.set_current_time(time);
+        self.data.set_current_time(time);
     }
 
     /// Returns a timer with the specified `id`.
     pub fn timer(&self, id: TimerId) -> Option<&TimerState> {
-        self.state.timers.get(id)
+        self.data.timers.get(id)
     }
 
     /// Enumerates all timers together with their states.
     pub fn timers(&self) -> impl Iterator<Item = (TimerId, &TimerState)> + '_ {
-        self.state.timers.iter()
+        self.data.timers.iter()
     }
 
     /// Iterates over pending [`WakeUpCause`]s.
     pub fn pending_wakeup_causes(&self) -> impl Iterator<Item = &WakeUpCause> + '_ {
-        self.state.waker_queue.iter().map(Wakers::cause)
+        self.data.waker_queue.iter().map(Wakers::cause)
     }
 
     /// Restores a workflow from the persisted state and the `spawner` defining the workflow.
@@ -311,23 +188,21 @@ impl PersistedWorkflow {
     ///
     /// Returns an error if the workflow definition from `module` and the `persisted` state
     /// do not match (e.g., differ in defined channels).
-    pub(crate) fn restore<'a>(
+    pub(crate) fn restore<S: CreateWorkflow>(
         self,
-        spawner: &WorkflowSpawner<()>,
-        services: Services<'a>,
-    ) -> anyhow::Result<Workflow<'a>> {
+        spawner: &WorkflowSpawner<S>,
+        services: Services,
+    ) -> anyhow::Result<Workflow<S::Spawned>> {
         let interface = spawner.interface();
         let data = self
-            .state
+            .data
             .restore(interface, services)
             .context("failed restoring workflow state")?;
-        let mut workflow = Workflow::from_state(spawner, data, self.args)?;
-        self.memory
-            .restore(&mut workflow)
-            .context("failed restoring workflow memory")?;
-        self.refs
-            .restore(&mut workflow.store)
-            .context("failed restoring workflow externrefs")?;
+
+        let engine_data =
+            serde_json::from_value(self.engine_data).context("failed deserializing engine data")?;
+        let mut workflow = Workflow::new(spawner, data, self.args)?;
+        workflow.inner.restore(engine_data)?;
         Ok(workflow)
     }
 }
